+++ CHANGES.md
@@ -1,4 +1,5 @@
 #6.0.28 ????-??-??
+ - 2020-03-02 Fixed bug#[14930](https://bugs.otrs.org/show_bug.cgi?id=14930) - ACL Match Setting 'Frontend' causes a problem with changing in Type, Queue, Service or SLA.
 
 #6.0.27 2020-03-20
  - 2020-02-28 Fixed bug#[14752](https://bugs.otrs.org/show_bug.cgi?id=14752) - The first valid certificate is preselected instead of the longest one.